Restructure resource content activities so menu only appears in viewers
not editors
